Finally figured out antialiasing + Util.bufferedImage
Use the flags in PathPlanner/README.md to enable antialiasing, System.setProperty doesn't seem to work, even with SwingUtilities.invokeLater
