+++ PathPlanner/README.md
@@ -1,0 +1,4 @@
+# PathPlanner & PurePursuit
+
+For both swing guis, use `-Dawt.useSystemAAFontSettings=on -Dswing.aatext=true -Dsun.java2d.xrender=true` in your JVM
+options to enable antialiasing.
